Division now uses floating-point numbers
Division now uses floating-point numbers instead of integers.

+++ calculator.py
@@ -12,7 +12,7 @@ if operator == "*":
     result = float(num1) * float(num2)
 #if division, then divide
 if operator == "/":
-    result = int(num1) / int(num2)
+    result = float(num1) / float(num2)
 #if addition, that add it together
 if operator == "+":
     result = float(num1) + float(num2)
